Move prompt to prompt.rs
