+++ src/prediction/multi_statement_classifier.rs
@@ -88,75 +88,6 @@ impl<R: MultiStatementResultTypes> MultiStatementPromptResult<R> {
     }
 }
 
-impl<R: MultiStatementResultTypes> MultiStatementPrompt<R> {
-    /// Gives a prompt that computes various meta information on the passed statements
-    // TODO: move to prompts.rs
-    pub fn statement_meta(stmts: &[Statement]) -> Self {
-        let mut stmts_s = String::from("");
-        for s in stmts {
-            stmts_s += format!("{}: {}", s.id, s.text).as_str();
-        }
-        MultiStatementPrompt {
-            name: "statement_meta".into(),
-            version: 5,
-            handler: |s| {
-                let s_without_header = s.trim().splitn(2, "\n").nth(1).unwrap_or("").to_string();
-                let target_delim_count = 7;
-                s_without_header
-                    .split("\n")
-                    // -- fixup delimiter count, since the ai does not reliably do that --
-                    .map(|s| -> String {
-                        let s = s.to_string();
-                        let delim_count = s.match_indices("|").collect::<Vec<_>>().len();
-                        if delim_count < target_delim_count {
-                            s + "|".repeat(target_delim_count - delim_count).as_str().into()
-                        } else if delim_count > target_delim_count {
-                            s.strip_suffix("|".repeat(delim_count - target_delim_count).as_str())
-                                .unwrap()
-                                .into()
-                        } else {
-                            s
-                        }
-                    })
-                    .collect::<Vec<_>>()
-                    .join("\n")
-                    // -- finally try to return an R (result type) --
-                    .try_into()
-                    .expect("Unable to extract from string")
-            },
-            primer: vec![
-                AiMessage::system(
-                    "
-You will be given multiple statements, each starting on their own line,
-and your task is to determine whether the statement falls into the category
-of politics or personal statements. In the case of it being a political category,
-give which political ideologies (e.g., liberalism, conservatism, socialism)
-each quote aligns with the most.
-In the case of it being a personal category, give the big five personality traits instead.
-
-In addition, also output up to three topic tags. The output should be a csv table with empty values as \"-\".
-All cells should be followed by a strength score (w=weak, s=strong) after a \":\" delimiter.
-If you are not sure, use \"-\"",
-                ),
-                AiMessage::user(
-                    "
-1. The global economy is at risk of recession due to the trade war and uncertainty it creates.
-2. In clubs kann man hervorragend neue Freunde kennenlernen",
-                ),
-                AiMessage::assistant(
-                    "
-num|category|label1|label2|label3|tag1|tag2|tag3
-1|politics|neoliberalism:s|conservatism:w|socialism:w|global economy:s|trade war:s|uncertainty:s
-2|personal|extraversion:s|openness:w|agreeableness:s|clubs:s|friendship:s|socializing:w
-",
-                ),
-                AiMessage::user(format!("{}", stmts_s).as_str()),
-            ],
-            stmts: stmts.to_vec(),
-        }
-    }
-}
-
 /// Used to generate prompts and handle the result
 ///
 /// Will cache results in the db based on the given prompt
@@ -168,19 +99,6 @@ pub struct MultiStatementPromptGen<'a, R: MultiStatementResultTypes> {
     /// Used for database access to e.g. find next statements to run the prompt on
     pub pool: &'a SqlitePool,
 }
-
-// TODO: this should probably go into prompts.rs
-/// A single row of the result that we get back via the multi_statement_predictor
-// #[derive(Serialize, Deserialize, Clone, PartialEq, Eq)]
-// pub struct StatementMeta {
-//     pub category: String,
-//     pub label1: String,
-//     pub label2: String,
-//     pub label3: String,
-//     pub tag1: String,
-//     pub tag2: String,
-//     pub tag3: String,
-// }
 
 impl<R: MultiStatementResultTypes> AiPrompt for MultiStatementPrompt<R> {
     type PromptResult = MultiStatementPromptResult<R>;
